added focus to the second text field in goToPage, also changed the lunch Icon for the app, needs some resizing
